refactor(dashboard): extract message renderer + shortcut cheat-sheet builder
Move two large pure derivations out of App.tsx:
- the custom chat-message renderer (permission/question prompts, tool
  bubbles/groups, evaluator banner, stall/resume chips) into
  useMessageRenderer — same 14-entry deps array, per-branch JSX unchanged,
  renderer component imports relocated to the hook.
- the registry-driven keyboard cheat-sheet builder (#4412/#4432/#3748) into
  the pure buildShortcutEntries helper — still called per render by design.

Eight now-dead component/util imports pruned from App. No behavior change.

Part of #5560 (App.tsx Shell decomposition).
